Added adjustment to position which handles price changes
To have a price discounted you enter a negative value, to have a price raised, enter a positive value. The value entered is the percentage subtracted or added to the position bean total.

The adjusment is output in the PDF (letterhead, blank) right behind the position price sporting a smaller font-size.